+++ plugins/CoreHome/javascripts/dataTable.js
@@ -237,6 +237,7 @@ $.extend(DataTable.prototype, UIControl.prototype, {
         var ajaxRequest = new ajaxHelper();
 
         ajaxRequest.addParams(params, 'get');
+        ajaxRequest.withTokenInUrl();
 
         ajaxRequest.setCallback(
             function (response) {
@@ -1367,6 +1368,7 @@ $.extend(DataTable.prototype, UIControl.prototype, {
                 action: 'saveViewDataTableParameters',
                 report_id: reportId
             }, 'get');
+            ajaxRequest.withTokenInUrl();
             ajaxRequest.addParams({
                 parameters: JSON.stringify(parameters)
             }, 'post');